# doc/openAPI.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Baby Statistic API",
    "description": "REST API for tracking baby milk, sleep, pee, and poop events.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local dev" }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/served-milk": {
      "get": {
        "tags": ["Served Milk"],
        "summary": "List all served milk records",
        "parameters": [
          { "$ref": "#/components/parameters/From" },
          { "$ref": "#/components/parameters/To" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ServedMilk" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Served Milk"],
        "summary": "Create a served milk record",
        "description": "status must be FRIDGE or FREEZER. expiryDate is auto-set (FRIDGE +4 days, FREEZER +6 months). originalAmount is auto-set to amount.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ServedMilkPost" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServedMilk" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
...
    "/api/served-milk/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "get": {
        "tags": ["Served Milk"],
        "summary": "Get a served milk record by ID",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServedMilk" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["Served Milk"],
        "summary": "Update a served milk record",
        "description": "Returns 409 if the record status is already USED or EXPIRED.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ServedMilkPost" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ServedMilk" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      },
      "delete": {
        "tags": ["Served Milk"],
        "summary": "Delete a served milk record",
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/api/pee": {
      "get": {
        "tags": ["Pee"],
        "summary": "List all pee events",
        "parameters": [
          { "$ref": "#/components/parameters/From" },
          { "$ref": "#/components/parameters/To" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Event" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Pee"],
        "summary": "Log a pee event",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Event" }
              }
            }
          }
        }
      }
    },
    "/api/pee/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "get": {
        "tags": ["Pee"],
        "summary": "Get a pee event by ID",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Event" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["Pee"],
        "summary": "Update a pee event timestamp",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EventPut" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Event" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Pee"],
        "summary": "Delete a pee event",
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/poop": {
      "get": {
        "tags": ["Poop"],
        "summary": "List all poop events",
        "parameters": [
          { "$ref": "#/components/parameters/From" },
          { "$ref": "#/components/parameters/To" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Event" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Poop"],
        "summary": "Log a poop event",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Event" }
              }
            }
          }
        }
      }
    },
    "/api/poop/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "get": {
        "tags": ["Poop"],
        "summary": "Get a poop event by ID",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Event" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["Poop"],
        "summary": "Update a poop event timestamp",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EventPut" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Event" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Poop"],
        "summary": "Delete a poop event",
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/medicine": {
      "get": {
        "tags": ["Medicine"],
        "summary": "List active medicines with latest log",
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Medicine" } } } }
          }
        }
      },
      "post": {
        "tags": ["Medicine"],
        "summary": "Create a medicine",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MedicinePost" } } }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Medicine" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
...
    "/api/medicine/logs": {
      "get": {
        "tags": ["Medicine"],
        "summary": "List medicine log entries",
        "parameters": [
          { "$ref": "#/components/parameters/From" },
          { "$ref": "#/components/parameters/To" }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/MedicineLog" } } } }
          }
        }
      }
    },
    "/api/medicine/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "put": {
        "tags": ["Medicine"],
        "summary": "Update medicine name",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MedicinePost" } } }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Medicine" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Medicine"],
        "summary": "Soft-delete a medicine (logs retained)",
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/medicine/{id}/active": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "patch": {
        "tags": ["Medicine"],
        "summary": "Set active state of a medicine",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "isActive": { "type": "boolean" } }, "required": ["isActive"] } } }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Medicine" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/medicine/{id}/log": {
      "parameters": [{ "$ref": "#/components/parameters/Id" }],
      "post": {
        "tags": ["Medicine"],
        "summary": "Log a dose taken",
        "requestBody": {
          "required": false,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MedicineLogPost" } } }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MedicineLog" } } }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      },
      "From": {
        "name": "from",
        "in": "query",
        "required": false,
        "description": "Filter by created_at >= value (ISO datetime)",
        "schema": { "type": "string", "format": "date-time", "example": "2026-04-13T00:00:00" }
      },
      "To": {
        "name": "to",
        "in": "query",
        "required": false,
        "description": "Filter by created_at <= value (ISO datetime)",
        "schema": { "type": "string", "format": "date-time", "example": "2026-04-13T23:59:59" }
      }
    },
    "responses": {
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Conflict": {
        "description": "Conflict — record is in a terminal state",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        },
        "required": ["error"]
      },
...
      "Timestamps": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "createdAt"]
      },
      "ServedMilk": {
        "allOf": [
          { "$ref": "#/components/schemas/Timestamps" },
          {
            "type": "object",
            "properties": {
              "amount": { "type": "integer", "example": 80 },
              "originalAmount": { "type": "integer", "example": 80 },
              "status": { "type": "string", "enum": ["FRIDGE", "FREEZER", "USED", "EXPIRED"], "example": "FRIDGE" },
              "expiryDate": { "type": "string", "format": "date-time", "nullable": true, "example": "2026-04-18T10:00:00.000Z" }
            },
            "required": ["amount", "originalAmount", "status", "expiryDate"]
          }
        ]
      },
      "ServedMilkPost": {
        "type": "object",
        "properties": {
          "amount": { "type": "integer", "example": 80 },
          "status": { "type": "string", "enum": ["FRIDGE", "FREEZER"], "example": "FRIDGE" },
          "originalAmount": { "type": "integer", "example": 80 },
          "expiryDate": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "required": ["amount", "status"]
      },
...
      "Event": {
        "allOf": [
          { "$ref": "#/components/schemas/Timestamps" }
        ]
      },
      "EventPut": {
        "type": "object",
        "properties": {
          "createdAt": { "type": "string", "format": "date-time", "example": "2026-04-14T12:00:00" }
        }
      },
...
      "Medicine": {
        "type": "object",
        "properties": {
          "id":            { "type": "integer" },
          "name":          { "type": "string", "example": "Vitamin D" },
          "isActive":      { "type": "boolean" },
          "latestTakenAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt":     { "type": "string", "format": "date-time" }
        }
      },
      "MedicinePost": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "example": "Vitamin D" }
        },
        "required": ["name"]
      },
      "MedicineLog": {
        "type": "object",
        "properties": {
          "id":         { "type": "integer" },
          "medicineId": { "type": "integer" },
          "takenAt":    { "type": "string", "format": "date-time" },
          "createdAt":  { "type": "string", "format": "date-time" }
        }
      },
      "MedicineLogPost": {
        "type": "object",
        "properties": {
          "takenAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      }
    }
  }
}